Added Connector to attach listeners to eventdispatcher. Added ServiceStub to allow lazy dependency injection.
ServiceStub can be seen as a locator for services (contains a Container-instance and alias).
The connector features lazyloading of listeners, by using a servicestub (which is recognized in ConnectedListener::getListener() ).
The connector can be registered with the tag "kernel.event_connector" to ContainerAwareEventDispatcher.
Tag "kernel.event_connector" has an optional listener-attribute containing the service-id of a listener (servicestub will be created in the background).
Added connector setter/getter to EventDispatcher.